src/converter_lambda/requirements-lambda.txt
apple-music-python
spotipy